Use custom index to speed up searches with many results
Use the custom (id, topic_id, posted) index on fluxbb_posts to speed up
searches that return a lot of results. I also added the posted field to
the predefined posts_multi_idx index so the "show_user_posts" action is
fast too.

On a related note, a key component of fast searching was to reorder the
fluxbb_search_matches table by the word_id column. This table has a lot
of rows (~73 million) and we want to be able to filter based on word_id
and quickly get the respective post ids. Previously it had to seek over
a couple of GBs to fetch the post ids. Ordering by word_id allows us to
get a (word_id, post_id) index for free (would take ~2GB to create it).